Refactor, structured logging, signal handling.
