+++ run/idp-sql/main.py
@@ -13,205 +13,29 @@
 # limitations under the License.
 
 import datetime
-from functools import wraps
-import json
-import logging
-import os
-from typing import Any, Callable, Dict, Union
+import signal
+import sys
+from types import FrameType
 
-import firebase_admin
-from firebase_admin import auth  # noqa: F401
+
+import database
 from flask import Flask, render_template, request, Response
-from google.cloud import secretmanager
-import sqlalchemy
+import middleware
+from middleware import jwt_authenticated, logger
 
-default_app = firebase_admin.initialize_app()
 app = Flask(__name__, static_folder="static", static_url_path="")
-
-logger = logging.getLogger()
-
-
-def init_connection_engine() -> Dict[str, int]:
-    db_config = {
-        # Pool size is the maximum number of permanent connections to keep.
-        "pool_size": 5,
-        # Temporarily exceeds the set pool_size if no connections are available.
-        "max_overflow": 2,
-        # The total number of concurrent connections for your application will be
-        # a total of pool_size and max_overflow.
-        # SQLAlchemy automatically uses delays between failed connection attempts,
-        # but provides no arguments for configuration.
-        # 'pool_timeout' is the maximum number of seconds to wait when retrieving a
-        # new connection from the pool. After the specified amount of time, an
-        # exception will be thrown.
-        "pool_timeout": 30,  # 30 seconds
-        # 'pool_recycle' is the maximum number of seconds a connection can persist.
-        # Connections that live longer than the specified amount of time will be
-        # reestablished
-        "pool_recycle": 1800,  # 30 minutes
-    }
-
-    if os.environ.get("DB_HOST"):
-        return init_tcp_connection_engine(db_config)
-    else:
-        return init_unix_connection_engine(db_config)
-
-
-# [START cloudrun_python_user_auth_secrets]
-def get_cred_config() -> Dict[str, str]:
-    if "CLOUD_SQL_CREDENTIALS_SECRET" in os.environ:
-        name = os.environ["CLOUD_SQL_CREDENTIALS_SECRET"]
-        client = secretmanager.SecretManagerServiceClient()
-        response = client.access_secret_version(request={"name": name})
-        return json.loads(response.payload.data.decode("UTF-8"))
-    # [END cloudrun_python_user_auth_secrets]
-    else:
-        logger.info(
-            "CLOUD_SQL_CREDENTIALS_SECRET env var not set. Defaulting to environment variables."
-        )
-        if "DB_USER" not in os.environ:
-            raise Exception("DB_USER needs to be set.")
-
-        if "DB_PASSWORD" not in os.environ:
-            raise Exception("DB_PASSWORD needs to be set.")
-
-        if "DB_NAME" not in os.environ:
-            raise Exception("DB_NAME needs to be set.")
-
-        if "CLOUD_SQL_CONNECTION_NAME" not in os.environ:
-            raise Exception("CLOUD_SQL_CONNECTION_NAME needs to be set.")
-
-        return {
-            "DB_USER": os.environ["DB_USER"],
-            "DB_PASSWORD": os.environ["DB_PASSWORD"],
-            "DB_NAME": os.environ["DB_NAME"],
-            "DB_HOST": os.environ.get("DB_HOST", None),
-            "CLOUD_SQL_CONNECTION_NAME": os.environ["CLOUD_SQL_CONNECTION_NAME"],
-        }
-
-
-def init_tcp_connection_engine(
-    db_config: Dict[str, str]
-) -> sqlalchemy.engine.base.Engine:
-    creds = get_cred_config()
-    db_user = creds["DB_USER"]
-    db_pass = creds["DB_PASSWORD"]
-    db_name = creds["DB_NAME"]
-    db_host = creds["DB_HOST"]
-
-    # Extract host and port from db_host
-    host_args = db_host.split(":")
-    db_hostname, db_port = host_args[0], int(host_args[1])
-
-    pool = sqlalchemy.create_engine(
-        # Equivalent URL:
-        # postgres+pg8000://<db_user>:<db_pass>@<db_host>:<db_port>/<db_name>
-        sqlalchemy.engine.url.URL(
-            drivername="postgresql+pg8000",
-            username=db_user,  # e.g. "my-database-user"
-            password=db_pass,  # e.g. "my-database-password"
-            host=db_hostname,  # e.g. "127.0.0.1"
-            port=db_port,  # e.g. 5432
-            database=db_name,  # e.g. "my-database-name"
-        ),
-        **db_config,
-    )
-    pool.dialect.description_encoding = None
-    return pool
-
-
-# [START cloudrun_python_user_auth_sql_connect]
-def init_unix_connection_engine(
-    db_config: Dict[str, str]
-) -> sqlalchemy.engine.base.Engine:
-    creds = get_cred_config()
-    db_user = creds["DB_USER"]
-    db_pass = creds["DB_PASSWORD"]
-    db_name = creds["DB_NAME"]
-    db_socket_dir = creds.get("DB_SOCKET_DIR", "/cloudsql")
-    cloud_sql_connection_name = creds["CLOUD_SQL_CONNECTION_NAME"]
-
-    pool = sqlalchemy.create_engine(
-        # Equivalent URL:
-        # postgres+pg8000://<db_user>:<db_pass>@/<db_name>
-        #                         ?unix_sock=<socket_path>/<cloud_sql_instance_name>/.s.PGSQL.5432
-        sqlalchemy.engine.url.URL(
-            drivername="postgresql+pg8000",
-            username=db_user,  # e.g. "my-database-user"
-            password=db_pass,  # e.g. "my-database-password"
-            database=db_name,  # e.g. "my-database-name"
-            query={
-                "unix_sock": "{}/{}/.s.PGSQL.5432".format(
-                    db_socket_dir, cloud_sql_connection_name  # e.g. "/cloudsql"
-                )  # i.e "<PROJECT-NAME>:<INSTANCE-REGION>:<INSTANCE-NAME>"
-            },
-        ),
-        **db_config,
-    )
-    pool.dialect.description_encoding = None
-    return pool
-
-
-# [END cloudrun_python_user_auth_sql_connect]
-
-
-# This global variable is declared with a value of `None`, instead of calling
-# `init_connection_engine()` immediately, to simplify testing. In general, it
-# is safe to initialize your database connection pool when your script starts
-# -- there is no need to wait for the first request.
-db = None
 
 
 @app.before_first_request
-def create_tables() -> None:
-    global db
-    db = init_connection_engine()
-    # Create pet_votes table if it doesn't already exist
-    with db.connect() as conn:
-        result = conn.execute(
-            "CREATE TABLE IF NOT EXISTS pet_votes "
-            "( vote_id SERIAL NOT NULL, "
-            "time_cast timestamp NOT NULL, "
-            "candidate VARCHAR(6) NOT NULL, "
-            "uid VARCHAR(128) NOT NULL, "
-            "PRIMARY KEY (vote_id)"
-            ");"
-        )
-        logging.info(result)
+def create_table() -> None:
+    database.create_tables()
 
 
 @app.route("/", methods=["GET"])
 def index() -> str:
-    context = get_index_context()
-    return render_template("index.html", **context)
-
-
-def get_index_context() -> Dict[str, Union[int, str]]:
-    votes = []
-
-    with db.connect() as conn:
-        # Execute the query and fetch all results
-        recent_votes = conn.execute(
-            "SELECT candidate, time_cast FROM pet_votes "
-            "ORDER BY time_cast DESC LIMIT 5"
-        ).fetchall()
-        # Convert the results into a list of dicts representing votes
-        for row in recent_votes:
-            votes.append(
-                {
-                    "candidate": row[0],
-                    "time_cast": row[1],
-                }
-            )
-        stmt = sqlalchemy.text(
-            "SELECT COUNT(vote_id) FROM pet_votes WHERE candidate=:candidate"
-        )
-        # Count number of votes for cats
-        cats_result = conn.execute(stmt, candidate="CATS").fetchone()
-        cats_count = cats_result[0]
-        # Count number of votes for dogs
-        dogs_result = conn.execute(stmt, candidate="DOGS").fetchone()
-        dogs_count = dogs_result[0]
+    context = database.get_index_context()
+    cats_count = context["cats_count"]
+    dogs_count = context["dogs_count"]
 
     lead_team = ""
     vote_diff = 0
@@ -229,37 +53,9 @@ def get_index_context() -> Dict[str, Union[int, str]]:
     else:
         leader_message = "CATS and DOGS are evenly matched!"
 
-    return {
-        "dogs_count": dogs_count,
-        "recent_votes": votes,
-        "cats_count": cats_count,
-        "leader_message": leader_message,
-        "lead_team": lead_team,
-    }
-
-
-# [START cloudrun_python_user_auth_jwt]
-def jwt_authenticated(func: Callable[..., int]) -> Callable[..., int]:
-    @wraps(func)
-    def decorated_function(*args: Any, **kwargs: Any) -> Any:
-        header = request.headers.get("Authorization", None)
-        if header:
-            token = header.split(" ")[1]
-            try:
-                decoded_token = firebase_admin.auth.verify_id_token(token)
-            except Exception as e:
-                logger.exception(e)
-                return Response(status=403, response=f"Error with authentication: {e}")
-        else:
-            return Response(status=401)
-
-        request.uid = decoded_token["uid"]
-        return func(*args, **kwargs)
-
-    return decorated_function
-
-
-# [END cloudrun_python_user_auth_jwt]
+    context["leader_message"] = leader_message
+    context["lead_team"] = lead_team
+    return render_template("index.html", **context)
 
 
 @app.route("/", methods=["POST"])
@@ -271,19 +67,11 @@ def save_vote() -> Response:
     time_cast = datetime.datetime.utcnow()
     # Verify that the team is one of the allowed options
     if team != "CATS" and team != "DOGS":
-        logger.warning(team)
+        logger.warning(f"Invalid team: {team}")
         return Response(response="Invalid team specified.", status=400)
 
-    # Preparing a statement before hand can help protect against injections.
-    stmt = sqlalchemy.text(
-        "INSERT INTO pet_votes (time_cast, candidate, uid)"
-        " VALUES (:time_cast, :candidate, :uid)"
-    )
     try:
-        # Using a with statement ensures that the connection is always released
-        # back into the pool at the end of statement (even if an error occurs)
-        with db.connect() as conn:
-            conn.execute(stmt, time_cast=time_cast, candidate=team, uid=uid)
+        database.save_vote(team=team, uid=uid, time_cast=time_cast)
     except Exception as e:
         # If something goes wrong, handle the error in this section. This might
         # involve retrying or adjusting parameters depending on the situation.
@@ -299,6 +87,20 @@ def save_vote() -> Response:
         response="Vote successfully cast for '{}' at time {}!".format(team, time_cast),
     )
 
+
+# https://cloud.google.com/blog/topics/developers-practitioners/graceful-shutdowns-cloud-run-deep-dive
+def shutdown_handler(signal: int, frame: FrameType) -> None:
+    logger.info("Signal received, safely shutting down.")
+    database.shutdown()
+    middleware.logging_flush()
+    print("Exiting process.", flush=True)
+    sys.exit(0)
+
+
+signal.signal(signal.SIGINT, shutdown_handler)  # handles Ctrl-C locally
+signal.signal(
+    signal.SIGTERM, shutdown_handler
+)  # handles Cloud Run container termination
 
 if __name__ == "__main__":
     app.run(host="127.0.0.1", port=8080, debug=True)
