+++ run/idp-sql/requirements.txt
@@ -4,3 +4,5 @@ pg8000==1.16.5 # Do not upgrade >=1.6.6 until https://github.com/tlocke/pg8000/c
 gunicorn==20.0.4
 firebase-admin==4.5.1
 google-cloud-secret-manager==2.1.0
+#python-json-logger==2.0.1
+structlog==20.2.0
